refactor: privatization
