+++ cognite_toolkit/_cdf_tk/load/_resource_loaders.py
@@ -118,7 +118,7 @@ from cognite.client.exceptions import CogniteAPIError, CogniteDuplicatedError, C
 from cognite.client.utils.useful_types import SequenceNotStr
 from rich import print
 
-from cognite_toolkit.cdf_tk.utils import (
+from cognite_toolkit._cdf_tk.utils import (
     CDFToolConfig,
     calculate_directory_hash,
     get_oneshot_session,
