+++ cognite_toolkit/_cdf_tk/templates/data_classes/_base.py
@@ -8,8 +8,8 @@ from typing import Any, TypeVar
 from rich import print
 
 from cognite_toolkit import _version
-from cognite_toolkit.cdf_tk.templates import BUILD_ENVIRONMENT_FILE
-from cognite_toolkit.cdf_tk.utils import read_yaml_file
+from cognite_toolkit._cdf_tk.templates import BUILD_ENVIRONMENT_FILE
+from cognite_toolkit._cdf_tk.utils import read_yaml_file
 
 
 @dataclass
