+++ cognite_toolkit/_cdf_tk/templates/data_classes/_config_yaml.py
@@ -13,15 +13,15 @@ from typing import Any, Literal, cast
 import yaml
 from rich import print
 
-from cognite_toolkit._version import __version__
-from cognite_toolkit.cdf_tk.load import LOADER_BY_FOLDER_NAME
-from cognite_toolkit.cdf_tk.templates._constants import (
+from cognite_toolkit._cdf_tk.load import LOADER_BY_FOLDER_NAME
+from cognite_toolkit._cdf_tk.templates._constants import (
     BUILD_ENVIRONMENT_FILE,
     DEFAULT_CONFIG_FILE,
     SEARCH_VARIABLES_SUFFIX,
 )
-from cognite_toolkit.cdf_tk.templates._utils import flatten_dict
-from cognite_toolkit.cdf_tk.utils import YAMLComment, YAMLWithComments
+from cognite_toolkit._cdf_tk.templates._utils import flatten_dict
+from cognite_toolkit._cdf_tk.utils import YAMLComment, YAMLWithComments
+from cognite_toolkit._version import __version__
 
 from ._base import ConfigCore, _load_version_variable
