+++ cognite_toolkit/_cdf_tk/templates/data_classes/_system_yaml.py
@@ -6,7 +6,7 @@ from typing import Any, ClassVar
 
 from rich import print
 
-from cognite_toolkit.cdf_tk.templates.data_classes._base import ConfigCore, _load_version_variable
+from cognite_toolkit._cdf_tk.templates.data_classes._base import ConfigCore, _load_version_variable
 
 
 @dataclass
