+++ tests/tests_unit/test_approval_modules.py
@@ -14,9 +14,9 @@ import pytest
 import typer
 from pytest import MonkeyPatch
 
-from cognite_toolkit.cdf import build, clean, deploy, main_init
-from cognite_toolkit.cdf_tk.templates import COGNITE_MODULES, iterate_modules
-from cognite_toolkit.cdf_tk.utils import CDFToolConfig
+from cognite_toolkit._cdf import build, clean, deploy, main_init
+from cognite_toolkit._cdf_tk.templates import COGNITE_MODULES, iterate_modules
+from cognite_toolkit._cdf_tk.utils import CDFToolConfig
 from tests.constants import REPO_ROOT
 from tests.tests_unit.approval_client import ApprovalCogniteClient
 from tests.tests_unit.utils import mock_read_yaml_file
